Refactor host metrics
