+++ app/src/main/java/org/astraea/app/metrics/KafkaMetrics.java
@@ -24,8 +24,6 @@ import java.util.stream.Collectors;
 import org.astraea.app.metrics.broker.BrokerTopicMetricsResult;
 import org.astraea.app.metrics.broker.HasValue;
 import org.astraea.app.metrics.broker.TotalTimeMs;
-import org.astraea.app.metrics.platform.JvmMemory;
-import org.astraea.app.metrics.platform.OperatingSystemInfo;
 import org.astraea.app.metrics.producer.HasProducerNodeMetrics;
 import org.astraea.app.metrics.producer.HasProducerTopicMetrics;
 
@@ -209,33 +207,6 @@ public final class KafkaMetrics {
           .stream()
           .map(HasValue::of)
           .collect(Collectors.toUnmodifiableList());
-    }
-  }
-
-  public static final class Host {
-
-    private Host() {}
-
-    /**
-     * retrieve broker's operating system info
-     *
-     * @param mBeanClient a {@link MBeanClient} instance connect to specific kafka broker
-     * @return a {@link OperatingSystemInfo} describe broker's os info (arch, processors, memory,
-     *     cpu loading ...)
-     */
-    public static OperatingSystemInfo operatingSystem(MBeanClient mBeanClient) {
-      return new OperatingSystemInfo(
-          mBeanClient.queryBean(
-              BeanQuery.builder()
-                  .domainName("java.lang")
-                  .property("type", "OperatingSystem")
-                  .build()));
-    }
-
-    public static JvmMemory jvmMemory(MBeanClient mBeanClient) {
-      return new JvmMemory(
-          mBeanClient.queryBean(
-              BeanQuery.builder().domainName("java.lang").property("type", "Memory").build()));
     }
   }
